Split the browser and its rules into files about one thing each
Three files were over the size the repo asks for, and two of them mixed pure
markup-reading with the stateful shell around it.

`test-browser.ts` keeps the browser itself — cookies, redirects, and the
requests it drives — and hands its reading to two new neighbours:
`parsing.ts` for a page's words and links, `forms.ts` for its forms and what
pressing a button would send. Both are markup in, answers out, so they can be
checked without a request. Callers that wanted a reader now take it from the
file that owns it rather than through the browser.

The rules' own suite is split the same way, by what each part is about:
sending a value, reading ticks, and filling a form in. Every test came across
unchanged; none was dropped.
